This assert appears to serve no purpose except to cause Inkscape to crash when the user chooses to exit the app when the XML editor is open. Commenting out the assert doesn't seem to cause any particular untoward behavior.   I've added the following todo in the code:
TODO: This assert is getting hit when the XML editor is open and the
user cancels an application quit (see Inkscape bug #864791).  But this
assert doesn't seem to provide any value, since this routine doesn't
refer to views at all.  So this should be reviewed in more depth to see
if commenting out the assert introduces any new undesired behavior.

(bzr r1390)