* src/main.cpp:
Added a command-line arguement --extension-directory to find what
extension directory Inkscape is configured for.  This should be
helpful for external extensions that want to use the same path
as Inkscape itself on their install.

* src/extension/extension.cpp, src/extension/init.cpp:

Putting in half of Aaron's fix, mostly the one to keep things from
crashing, next we need to examine how to add multiple search paths
for extension dependencies.

(bzr r6545)