Reimplemented Inkboard's listeners at the undo log level instead of the node observer level. Currently Inkboard is using a very naive undo/redo model: commits to the undo log are propagated to all clients, and all clients push that log onto their own undo stacks.
This has severe limitations when it comes to maintaining consistency, but it's a place to
start implementing more sophisticated, safe algorithms/heurisitics for maintaining
document consistency.

Changes are no longer immediately committed.  Instead, Inkboard now queues up changes from
each receiver, and only commits a set of changes from a specific client when a CHANGE_COMMIT
message is sent.  This breaks compatibility with the old Inkboard CVS, but the protocol
has not been reversioned since this code has not yet had a public release (in the traditional
sense, anyway.)

Commits do not yet specify ancestors, but this should not be hard to
implement in the new framework.

(bzr r7527)