sorta-kinda fixed another bug with Inkboard and the text tool...
Judging from serialized messages and tracing the reconstruction of the document
tree on the receiving end, it seemed that the root cause of all these bugs
was subtrees that had their parents switched around via removal and later
re-addition to a tree.

The deserialization class dealt only on the level of adding and removing individual
nodes; it did not consider this possibility.  Now, when a serialized addChild event
is received, Inkboard looks in its buffer of recently added nodes.  If a node matches
an incoming serialized addChild event, Inkboard will use the node in the buffer
rather than creating a new one.  This allows us to move subtrees between parents without
explicit reconstruction.

This fixes _some_ bugs, but not all, which is why I wrote "seemed that the root cause
was..." above.  It seemed plausible when I was tracing the scenario out, but I guess
that wasn't all.

I'm committing this fix anyway.  Hopefully I'll find the real problem tomorrow
or sometime.

(bzr r7554)