Cleanup:   Consistant variable names.   x, y used for positions.   i, j used for column/row.
Added check boxes to remove the automatic adding of tile width and height
in calculating shift. Useful when placing clones on a circle or spiral
(and in some cases when using the Exponent parameter where adding a
-100% shift is not the same as excluding the shift).

Added "Base" parameters to scaling. This allows drawing of logarithmic
spirals such as found in nature (e.g. nautilus shell, pine cone). When
non-zero, it changes the scale factor to be base^(scale-1).

Added "Cumulate" to shift, scale, rotate. In the case of shifting,
it allows uniformly scaled clones to be placed touching each other.
For example, a "10% shift" normally results in clones being shifted
by 10%, 20%, 30%, ... With the Cumulate box checked, the shifts
become 10%, 10+20%, 10+20+30%, ... Adding "Cumulate" to the Scale and
Rotation tabs may be of dubious value... but it can easily be removed.

(bzr r12989)