Patch from Marcin that fixes bitmap effects where translation is not being used properly.
[Marcin Floryan]
I found the culprit and created a patch (attached). I hesitated for a
moment as to where to fix the problem, since there are two ways of doing
it (in paramenum.cpp and in the extensions files). Since I don't know
the reasons behind the particular line in the former I fixed the
latter. Should help.

PS. I have a feeling more could be fixed as it seems translations are
ignored in some other places for internal (cpp) extensions, but I can
see into that as a next step.

(bzr r13545)