implement UI for polygonal grid
Modify the UI for polygonal grid to have the appropriate parameters,
rather than those copied from axonometric grid. Since the grid logic has
not been changed yet, we are using the rotation angle as both angles for
the axonometric logic (which gives a degenerate grid with the default
value 0).

(bzr r11939)