Rename --do-not-fix-pre-92 to --no-convert-text-baseline-spacing
We will be introducing additional legacy file conversion options
(e.g. dpi conversion), which need to be controllable independently of
this baseline fix, so rename this command line option to more clearly
communicate what functionality it provides.

The terminology "convert" is chosen as a jargon word to be used for all
such legacy file conversions.  The "--no-XXX" naming style is adopted
from the convention used by other software such as GIMP.

Also add a man page entry for the option.

(bzr r15350)