document: Split out font name conversion separate from baseline fix
r15338 introduced a fix to replace 'Sans' font names by 'san-serif'
which is more widely supported in web browsers.  This fix is distinct
from the text baseline spacing fix and so should be applied
independently from that fix.

This conversion is being applied unconditionally because we no of no use
cases where one would wish to preserve the old font naming.  If it turns
out there is a reason why one would wish to disable it, it is
recommended this be implemented as a command line option called
"--no-convert-font-names".

(bzr r15351)