Make file update dialog transient for the main window
This ensures the (modal) dialog is positioned on top of the Inkscape
window in which the file will be opened.
Before it could happen that the dialog was not even on the same
monitor as the Inkscape window leaving Inkscape in a seemingly
inaccessible state if it was overlooked.

(cherry picked from commit 5981085fa98a7cc785524f89f107cc7ff43c8b42)