Prevent setting zero radius
While the spec explicitly allows for that (and it should disable
rendering of the element) Inkscape does not allow setting a zero
radius at runtime (it works fine when opening an existing file).

If we still force this during runtime we end up with a "ghost"
ellipse on canvas which looks as if it had the old radius but has
control points according to the zero radius and has the added bonus
of not being movable (or rather being movable but snapping back
back immediately when released) until updating the radius with a
valid number again which finally moves the ellipse to wherever it
should have been.

(cherry picked from commit dfad76343452004bb06c92d19262d22dc3cb9e76)