This solves the jerky redraw when dragging objects (or no redraw when dragging quickly).
There are 4 events involved for dragging it seems:
1. GDKEventMouse
2. selection modified
   generated in Selection::_scheduled_modified() with priority 101
3. redraw
   generated in SPCanvas::addIdle() with priority 200
   calls SPCanvas::paint() which only redraws to *offscreen* buffer
4. refresh/expose generated in gtk_widget_queue_draw_area() with priority 120
   calls SPCanvas::handle_draw() which actually updates the screen

For fast redraw response, you want to execute #2, then #3, then #4 as soon as possible. This
requires the priorities for each event to be higher than the previous or else mouse events that
come later will get processed before earlier redraw,expose events already in the queue. Since
this is real time stuff, the longer an event waits in the queue, the higher its priority should be,
so this is a priority inversion.

What I don't understand is:
1. why this problem didn't happen in 0.92 (GTK2)

2. why separate redraw and expose, which adds latency
  It seems it's to avoid tearing since redraw can take longer than the monitor refresh interval
  But why not use double buffering and swap buffers immediately after redraw - that seems faster
  than generating an expose event