Automate buildsystem setup for macOS
The new 'install_toolset.sh' is now used as 'before_script' in CI
to setup the build system. It is now capable to install and switch
between sets of build systems by itself, no manual setup required
on the runner. This way a branch can be developed using different
versions of dependencies than another, and CI is able to switch
back and forth between them as needed.

The build system comes now as a compressed disk image. It's mounted
read-only and union-mounted with a writable ramdisk as overlay.
This greatly reduces the requirements for disk space and memory
usage while keeping optimal speed (and not wearing out the SSD).