Fix crash when xml editor is open and text is deleted
GtkTreeView has selection and cursor. Selection is what appears selected
and emits `changed` events. Cursor is another type of selected node,
that follows what the user pressed last, and is used for keyboard
navigation (like pressing up/down keys). When the selection is set
programatically the cursor is not changed. To observe this behavior:
* create 2 rectangles
* open the XML editor, and select one of them via the editor
* then using the select tool select the other rectangle (in the
document). This automatically changes the selected node in the XML
editor (but not the cursor!)
* now return back to the XML document and press the UP arrow
* you will notice that the newly selected row is not UP from the
previous selection, it is up from the previous cursor (which was the one
selected via the mouse)

So now to issue #619.

When a text node's content is selected in the XML editor, and then the
text is double-clicked and deleted via the document, the cursor points to the content, and
the selection points to the text node (there is a span in between).

    svg:text  <-- selection is here
      svg:tspan
        "text here" <-- cursor is here

The deletion causes the text to be deleted, which causes tspan to be
selected, because there is automatic cursor moving from the deleted node
up. Automatically selecting the tspan in the xml editor causes the tspan
to be selected in the document, which for some reason crashes, because
it's empty? I haven't figured out why this crashes, but I figured it is
not supposed to be selected.

So instead, when the text node is double clicked we can also move the
cursor to the svg:text element, and then the deletion of the contents
won't cause the svg:tspan to be selected, and there will be no crash.

In general it makes sense for the cursor to track the selection. Then
the keyboard shortcuts (up/down) will move with the selection, and there
won't be the situation described in the begining.

Fixes #619.

NOTE: It might still be practical to understand why selecting an empty
svg:tspan causes a crash. The top of the stacktraces provided in #619
list the offending calls.


(cherry picked from commit 81a2c47bb19b440015a18a437ff7651fff8d241a)