Make the large polka dots pattern repeat seamless
Some of these dots extended outside the pattern bounding box and were
clipped off. This  commit duplicates those dots in on the other side of
the pattern, so the two partial dots together form a whole dot again.

One dot was already repeated in this way. Another dot was so close to
the edge that it was moved 0.003 instead of repeated.

The medium and small polka dot patterns did not have this problem
because the dots are smaller, but since the dot positions were identical
between all sizes, the one dot that was moved is also moved in the
medium and small patterns to preserve this consistency.

An alternative fix could be to prevent the clipping in the first place
(which could have the advantage that the fill would produce complete
circles rather than some half circles). It seems this would be done by
overriding the default `overflow: hidden` CSS rule for patterns, but
setting that to `overflow: visible` produces undefined rendering
behavior according to the SVG specification 1.1 and 2.0, though 2.0 has
a recommendation). Also, trying this shows Inkscape still seems to apply
the clipping even with `style="overflow: visible !important;"`.

This commit was tested in Inkscape 1.0.2 by copying the file there,
but I assume the fix works identical in newer versions.