Fix SVG embedding
Modify SPUse to use the referenced object's document when rebuilding its
tree, rather than its own. For "regular" clones, these will be the same,
but this is critical for correct rendering when embedding one SVG into
another and is, more or less (see below), the correct behavior according
to the specification.

Also modify SPDocument::getObjectById to walk up the document tree
looking for the object, as this is needed for the embedded SVG to be
able to resolve references against the embedding document if they cannot
be resolved "internally". (So far as I can tell, this is the correct
behavior according to the specification.)

As I understand the specification, this implementation is technically
non-conforming, because it will resolve references *anywhere* in the
embedded document, not just within the selected subtree. However, a)
from my (admittedly limited) understanding of the code, I believe it is
somewhat challenging to implement this strictly according to the spec,
and would meanwhile like it to be less broken, and b) I suspect this
behavior is actually more *useful* anyway.

This does cause a problem with application of CSS rules, in that the
embedding document's rules are no longer applied. This will be fixed in
a subsequent commit.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1757.