Move unSymbol from selection code to symbol code
The unSymbol codes unlinks a symbol which is usually an object hidden in defs and unselectable.
The way unSymbol used to work forced the selection of an unselectable object from the symbols dialog and acted on it...

This also converts the verb (currently GUI-inaccessible) into unSymboling all selected uses of a symbol.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2437