Don't always set snap tolerances in document
Remove logic to always set the snap tolerances when creating a new named
view. This restores the behavior (and default snap settings) from
Inkscape 1.0.

Note that there is some debate whether snap tolerances should EVER be
saved in SVG, as they are likely to be dependent on the user and the
user's environment (screen size, in particular), more so than the
document.

See also https://gitlab.com/inkscape/inbox/-/issues/4852.