Fix overwriting of custom shortcuts
The primary shortcut (which is displayed in menus)
shouldn't be removed when removing another shortcut.

This checks that the removed shortcut is the same as the
primary shortcut, before removing the primary shortcut
attached to the verb.

Fix https://gitlab.com/inkscape/inkscape/-/issues/2384
Fix https://gitlab.com/inkscape/inkscape/-/issues/2453