Fix issue with export filenames not being saved
The original checks which should be checking to see if the
filename is based on the id or not, is clearly nonsense.
find_last_of doesn't do what the code thinks it does.

If we want to not save the filename sometimes, we should
write better code to handle those checks. For now, removing.