Update macOS build pipeline
This commit concentrates on build system improvements. Most of our
dependencies are being updated, laying the groundwork for more
functional improvements in the next update.

Most notably we are raising the minimum OS requirement to High
Sierra 10.13, a step which we put some effort into to successfully
avert it for Inkscape 1.1. But sadly, updating libraries causes
new issues on 10.11, making it too burdensome to support.

Backward compatibility of the Python runtime and packages have
been improved by reworking the external projects that provide
them. We also build and use wheels for Python packages now, the
last missing piece to not require downloads during CI.