Save doubles to preferences with 15 digits
Standard behavior of Glib::ustring::compose uses 6 significant
digits, which might be unnecessarily limiting in some cases.

At the same time the use of std::numeric_limits<double>::digits10
ensures that we only write precise digits, so that numbers that can't
be represented precisely in binary are rounded of nicely.