Fix performance issue of zooming
All objects, which include the invisible, are updated when we zoom
in/out, the improper algorithm leads to zooming lag.
The lag is proportional to the number of objects we create, and it
becomes unacceptable (on MacOS) when there are more than 1000 objects.

We should update what is visible while zooming and Skipping updating the
invisible objects when we take zoom actions to address the zooming problem.

This fix also address [issue #1614](https://gitlab.com/inkscape/inkscape/-/issues/1614)