Fixed a bug where the button press event was being blocked if both left and right click buttons are pressed.
As far as I understood, the "block_button" function was created so that
the popup menu doesn't appear when right click is pressed while a a left
click is being held. this is for example to prevent the popup menu from
appearing when someone is dragging an object or while dragging anything
else. This was happening but not as intended. This was also blocking
any event while both left and right click was clicked. this happens at
the part that checks for event_context->block_button(event) and return
(block the event) if the return value is true (if both buttons are being
pressed). The current solution is to not to use the return value for
now, just set what buttons are being pressed, and later check for it
before taking actions (like before showing the popup menu).

Some tools use the right click. The selector tool for example, uses
right click to abort the selection proccess for both the normal
selection mode and for the rubberband mode. The part where this code
lies were never executed due to this bug.

This fixes lots of tools being stuck while both buttons were being
pressed. The tools include:
 - Selector tool
 - Rectangle tool
 - Ellipse tool
 - Star tool
 - 3D Box tool
 - Spiral tool
 - Pen tool
 - Calligraphy tool
 - Text tool
 - Mesh tool
 - Dropper tool
 - Paint Bucket tool
 - Tweak tool
 - Spray tool
 - Eraser tool
 - Connector tool
 - Zoom tool
 - Measure tool

Tools that still suffer from this problem:
 - Pencil tool