Fixed the right click bug for the Pencil tool.
The writer of the code wanted not to keep drawing when the buttons that
are responsible for scrolling are being pressed. button 3 (right click)
isn't used for scrolling and the drawing should continue unless some
other logic (like abort the drawing) is implemented for the right click
button in the future.