Set the pen tool to ignore right clicks when the left click is being pressed.
The pen tool had 2 problems:

 1 - right click finishes the creation of the paths even if the user is
still dragging. This is solved by just checking whether the left click
button is pressed or not. If not, then finish. Otherwise, ignore the
right click event.

 2 - The dragging (the motion event) was being blocked for both the
scroll wheel button and the right click button. This has caused the
dragging to stop if right click was being pressed. It only makes sence
to block if the scroll wheel button is being pressed. The solution here
is to just remove the right click button from the set of buttons that
blocks the motion event.