Added one more option to SPDocument::getItemsAtPoints to allow for the case of having multiple items overlapping at a point.
SPDocument::getItemsAtPoints was only returning the topmost items. If
there are two items that intersect with the same point, it'll return
only the top item. Now, this is an option called "topmost_only". if
topmost_only is false, all items that are at a same point will be
returned as well.