Ignore several text position/shape propteries in pasteStyle
The past style feature is intended to paste the look
of an object, but css now includes some parts which
are not really related to the style, but to the layout
of the text. These should be stripped out.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/2807
(cherry picked from commit 4bd7e42e83568cf80f3799e73635126c17c1a956)