Add macOS shared runner
macOS shared runners aren't publicly available yet, we are
participating in a beta access program. The shared runner runs its
own job in parallel to our self-hosted runner. It does not interfere
or affect our releasing; our self-hosted runner keeps its status
"official Inkscape CI". (This might change in the future.)

Building on Catalina using the 10.13 SDK requires two jobs in our
pipeline and the (temporary) introduction of a "prepare" stage.
Pending further discussions with GitLab, this can be considered a
workaround and will be changed in the future.