Moves keyboard signal exception to SPDesktop
Most of the time, when the mouse is over the canvas, a keyboard
signal can be sent back up the stack to the event context.

There was older code that tried to do this in the canvas grid,
but this is far too late for the main context.

The extra checks as to make sure this exception is limited as possible.