Remove NSIS License agreement from windows install
The License agreement screen in the NSIS installer is
created by CPACK, this doesn't provide us with any
control over the presentation of the GPL to the user
and they are asked to "agree" to it. This is incorrect.

Now that we have a Welcome screen, we can remove this
as we can always choose to show the GPL to new users at
any time, which will be more consistant than this.