Fix problems with toolbo prefs by splitting widgets
the preferences keep breaking when the toolbox is
modified. This is because they share the same ui
file and try to modify it.

This split requires future additions to the toolbox
to modify both of these files. But are much better
about defining what they show to users and need
much less messing about.