Fix #1557 - SEGFAULT when Flood tool busy working
With the Flood tool busy working (either because the use spam-clicked
with it or because the operation is slow due to a high "Close Gaps"
setting), quickly switching tools used to cause a cause. This commit
provides a workaround preventing the crash.

The root cause of the problem is incorrect lifetime management of tools,
whereby the current desktop can delete the tool while the tool's member
function is still running in another thread. After control returns to
the member function, its `this` pointer is junk and any attempt to
access members is UB and will often segfault.

The temporary fix (before a better mechanism is implemented) is to copy
all the necessary data onto the stack before yielding to the other
thread, so that we never dereference the dangling `this` pointer.
The analysis done by Tavmjong Bah on the bugtracker at issue #1557 was
very helpful in the development of this fix.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1557
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3137