Improve appearance, UX aspects of guide handles
Guide handles are the origin controls that used to show as red "dots"
along the on-canvas guides. Until now, their only purpose was marking
the rotation origin of a guide. This commit makes the guide handles
interactive and improves their visual appearance.

The size of the guide handles is now adjustable in the Preferences via
the existing setting Interface - Handle size. However, the size of the
guide handles scales more slowly than the size of other controls and is
effectively equal approximately half the size of an auto-smooth node.
At the same time, a minimum size is enforced, equal to the previously
hard-coded size.

As a result, the guide handles start enlarging when the "Handle Size"
preference is set to a value of 7 or larger, addressing the needs of
users on ultra-high DPI displays as well as of users with physical or
visual impairment. Since the default value of this setting is 3, the
size of the handles is unaffected on new installs or for users on
default settings.

A handle now has a hitbox which makes it easier to interact with the
guide. This is implemented by forwarding the handle's events to the
existing event handler of the line.

The color of the handle is now the same as the color of the guide,
instead of the hard-coded red. When a guide is locked, its handle
always becomes gray, in addition to turning into an "x" shape (as
before).

A few small fix-ups were done to the surrounding code along the way:

* returning on error instead of proceeding to dereference nullptr;
* removal of some unused #include's;
* corrected whitespace in a few places.

Many thanks are due to the people who helped review and refine the UX
aspects of these changes, especially Adam Belis, Maren Hachmann, and
Mike Kowalski.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3111