Rework pixbuf_to_png to fix multiple bugs
The first main issue is that the implementation makes all manner of
unaligned accesses, which is UB. For 16-bit RGBA, writing 0 to ptr as a
64-bit integer for the !pad is legal, but for every other combination at
least some of the time it will be unaligned. Moreover, this write will
access outside the bounds of the image data when on the last one or few
(depending on color_type) pixels of the image, and is not even necessary
in most cases due to the (in-bounds) data being fully overwritten later,
though this "out-of-bounds" access is currently not itself a problem due
to the + 64 in the call to malloc, discussed more later. For RGB it
treats the pointer as a guint32*/guint64* despite only wanting to write
3 samples, leaving the high sample in memory as 0. On the next iteration
this high sample will be overwritten by ptr+3 (or ptr+6 for 16-bit),
forming a pattern of overlapping word accesses, 3/4 of which will be
unaligned. For <16-bit greyscale it still uses the pointer as a guint16*
unnecessarily, as it's only ever writing a maximum of 8 bits each time,
and for everything except 8-bit greyscale with alpha that will be
unaligned some of the time.

The second issue is the calculation of the size of buffer to allocate.
There is an undocumented + 64 added to it that has been present ever
since this function was added, and does not make any sense. This is more
than enough to compensate for the out-of-bounds accesses described
above, which will only be a maximum of 7 bytes past then end, far short
of 64. However, the calculation completely fails to take into account
that, for formats where the number of bits per pixel is less than 8,
there will be padding on every row; the current calculation assumes that
the pixels will be packed tightly with no padding. This means that if
n_fields * bit_depth * num_cols is not a multiple of 8 (for example,
1-bit greyscale without transparency with num_cols % 8 == 1, which is
the worst case that happens), padding slowly accumulates and, with
enough rows, the number of padding bits will eventually exceed the 64*8
spare bits arbitrarily added to the allocation.

The third issue is that the big-endian implementation is half-baked at
best, missing support for many of the formats (not even RGBA), and the
greyscale implementation looks to be completely wrong endianness-wise,
among other things.

Thus, rework the entire implementation to fix these issues. The first
change is to accurately compute the number of bytes for the image by
calculating the rounded number of bytes per row and multiplying by the
number of rows, dropping the +64. The second change is to avoid all the
unaligned accesses by making sample-level accesses when pixels aren't a
power of two number of samples, only initialising in the !pad case when
we have <8-bit samples, and doing so just for that byte. The final
change is to lightly adapt this reworked little-endian implementation to
also support big-endian targets. Whilst here some style issues have also
been cleaned up.

With these changes, running the test suite in an armhf chroot on an
arm64 kernel now passes, which previous hit a few SIGBUSes due to the
unaligned accesses. This also fixes the following 36 test cases (with 49
failures remaining) on big-endian targets (tested on s390x):

    cli_export-type_png_check_output
    cli_export-area-snap_export-id_check_output
    cli_export-area-snap_export-area-drawing_check_output
    cli_export-width_check_output
    cli_export-width_export-dpi_check_output
    cli_export-height_check_output
    cli_export-height_export-dpi_check_output
    cli_export-id_png_check_output
    cli_export-id_export-id-only_png_check_output
    cli_export-id_export-id-only_export-area-page_png_check_output
    cli_export-with-filters_png_check_output
    cli_export-default-background_png_check_output
    cli_export-background_png_check_output
    cli_export-background-opacity_png_check_output
    cli_actions-file-open_check_output
    cli_actions-object-align_check_output
    cli_actions-object-distribute_check_output
    cli_actions-object-set-attribute_check_output
    cli_actions-path-simplify_check_output
    cli_actions-object-unlink-clones_check_output
    cli_actions-transform-remove_check_output
    cli_actions-transform-rotate_check_output
    cli_actions-transform-scale_check_output
    cli_actions-transform-translate_check_output
    cli_import_cdr2_check_output
    cli_import_vsd_check_output
    cli_import_vsdx_check_output
    cli_import_wpg_check_output
    cli_systemLanguage_en_check_output
    cli_systemLanguage_fr_check_output
    cli_systemLanguage_fr2_check_output
    cli_systemLanguage_de_check_output
    cli_systemLanguage_de-CH_check_output
    cli_systemLanguage_pt_check_output
    cli_systemLanguage_xy_check_output
    cli_systemLanguage_fr_RDF_check_output

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2821
See https://gitlab.com/inkscape/inkscape/-/issues/1365