Correct alignment of on-canvas controls
On-canvas controls include elements such as node handles and guide
handles. In order to ensure sharp appearance, the positioning of these
controls is aligned to pixels in a CanvasItemBuffer.

This commit corrects the calculation of the pixel position of the
controls by ensuring that the position is always rounded down with the
`floor()` function.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2022
Fixes https://gitlab.com/inkscape/inkscape/-/issues/2604