Fix snapping for out of viewport objects
In order to slightly improve the performance while snapping, instead
of looking at each and every point in the document, we only look for
points that are inside the viewport. How we do this is the reason for
this bug.

If top-left or bottom-right points of the BBox are inside the viewport
then we 'consider' the object to be inside the viewport and not if otherwise.
We should just check if the viewport intersects with the object BBox

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3174