Lifecycle + file preview fixes
Lifecycle and crash fixes

* Refactored canvas initialisation code.

  The canvas now defines an "active" state as one in which it has both a
drawing and is realized. Some persistent initialisation now happens on
construction, while other more transient initialisation now happens on
activation. Most code only runs in the active state.

  All signals and callbacks are only scheduled in the active state, and
cancelled upon leaving it, ensuring signals and callbacks only run in
the active state. This invariant is also verified by `assert`s.

  This split gives the canvas a well-defined lifecycle model, and should
eliminate crashes involving the canvas's lifecycle/destruction.

* Only connect callbacks when not already connected.

  Otherwise when we disconnect, we only disconnect the last one. The
first one may still be called. This fixes a lingering crash when mucking
around with the open file dialog and closing things. (Inb4 the testers
find it.)

File previewer

* Changed from a ScrolledWindow to a Bin, since the canvas doesn't need
help scrolling, and this also removes a non-functional scrollbar.

* Removed requests to redraw/repaint, since the canvas decides when and
what to redraw as necessary, and forcing it introduces unnecessary
slowdowns on resize.

* Fix incorrect affine calculation - the canvas ignores, and always has
ignored, the translational component of the _affine variable. The
translational part should be set using scroll_to (renamed to set_pos),
just as it's done in SPDesktop. (This used to work, but it was relying
on a bug.) This ensures the drawing is centered - before it was
somewhere off-screen, making it look like the previewer was broken.

Minor changes

* Removed unused _need_repick variable - wasn't read by anything, only
set.

* Renamed a few functions (scroll_to -> set_pos, canvas_item_clear ->
canvas_item_destructed) to better reflect their use.