Allow cutting through clones with the Eraser tool
The Eraser tool no longer ignores clones when it is used in the "Cut out
from paths and shapes" mode. When the Eraser encounters a clone (a USE
item), it will unlink it and cut-erase the resulting ordinary object in
the usual way.

Fixes https://gitlab.com/inkscape/inbox/-/issues/6121